+++ .eslintignore
@@ -1,2 +1,3 @@
 example/RNBackgroundExample/node_modules
 example/RNBackgroundExample/e2e
+example/server/node_modules
