+++ example/server/src/index.ts
@@ -9,11 +9,14 @@ router.get('/', (_, res) => {
 });
 
 router.post('/upload', (req, res) => {
-  const filePath = path.join(__dirname, `/image.jpg`);
+  const filePath = path.join(__dirname, '/image.jpg');
   const stream = fs.createWriteStream('/dev/null');
 
   if (req.query.simulateFailImmediately) {
-    res.status(500).send('Simulated Error').end();
+    res
+      .status(500)
+      .send('Simulated Error')
+      .end();
     return;
   }
 
@@ -32,7 +35,10 @@ router.post('/upload', (req, res) => {
 
       console.log(`Processing  ...  ${progress.toFixed(2)}%`);
       if (req.query.simulateFailMidway && progress > 50) {
-        res.status(500).send('Simulated Error').end();
+        res
+          .status(500)
+          .send('Simulated Error')
+          .end();
         resolve();
       }
     });
@@ -43,7 +49,7 @@ router.post('/upload', (req, res) => {
       resolve();
     });
 
-    stream.on('error', (err) => {
+    stream.on('error', err => {
       console.error(err);
       reject(err);
     });
