+++ lib/index.js
@@ -1,3 +1,12 @@
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
 var __rest = (this && this.__rest) || function (s, e) {
     var t = {};
     for (var p in s) if (Object.prototype.hasOwnProperty.call(s, p) && e.indexOf(p) < 0)
@@ -107,6 +116,12 @@ export const ios = {
     Each chunk file will be named by its corresponding index (0, 1, 2,...).
     */
     chunkFile: (parentFilePath, chunkDirPath, numChunks) => NativeModule.chunkFile(parentFilePath, chunkDirPath, numChunks),
+    /*
+    Directly check the state of a single upload task.
+    Note that there's no way to distinguish between a task being completed, errored, or non-existent.
+    They're all "inactive". You will need to check that with whichever API service you're using.
+    */
+    getUploadStatus: (jobId) => __awaiter(void 0, void 0, void 0, function* () { return (yield NativeModule.getUploadStatus(jobId)) || { state: 'inactive' }; }),
 };
 export default {
     startUpload,
