--- lib/src/types.js
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=types.js.map
