+++ lib/types.d.ts
@@ -13,14 +13,14 @@ export interface CompletedData extends EventData {
     responseBody: string;
 }
 export declare type UploadId = string;
-export declare type FileInfo = {
+export interface FileInfo {
     name: string;
     exists: boolean;
     size?: number;
     extension?: string;
     mimeType?: string;
-};
-export declare type NotificationOptions = {
+}
+export interface NotificationOptions {
     /**
      * Enable or diasable notifications. Works only on Android version < 8.0 Oreo. On Android versions >= 8.0 Oreo is required by Google's policy to display a notification when a background service run  { enabled: true }
      */
@@ -69,7 +69,7 @@ export declare type NotificationOptions = {
      * Sets notification cancelled message   { onCancelledMessage: "Video upload was cancelled" }
      */
     onCancelledMessage: string;
-};
+}
 export interface UploadOptions {
     url: string;
     path: string;
